requirements.txt
pandas
yfinance
ta
scikit-learn
alpaca-py
python-dotenv
fastapi
uvicorn
